remove scratch

+++ backhack/main.py
@@ -49,12 +49,11 @@ def getgraphs(df, model, non_numeric_features):
     
     df=df.drop(non_numeric_features, axis=1)
     
-    try:
-        file = open('files/scratch.png')
-    except IOError as e:
-        pass
-    else:
+    if os.path.isfile('files/scratch.png'):
+    # Удаляем файл, если он существует
         os.remove('files/scratch.png')
+    else:
+        pass
     # explain the model's predictions using SHAP
     # (same syntax works for LightGBM, CatBoost, `scikit-learn, transformers, Spark, etc.)
     explainer = shap.Explainer(model)
